+++ src/components/CustomButton.tsx
@@ -1,0 +1,42 @@
+import {styled} from "@mui/material";
+import Button from "@mui/material/Button";
+import React from "react";
+
+interface CustomButtonProps {
+  color: "primary" | "secondary";
+  variant: "contained" | "outlined";
+  size?: "small" | "medium" | "large";
+  name?: string;
+  onClick: (_event: React.MouseEvent<HTMLButtonElement>) => void;
+  children: React.ReactNode;
+}
+
+const CustomButton: React.FC<CustomButtonProps> = ({
+  color,
+  variant,
+  size = "small",
+  name,
+  onClick,
+  children,
+}) => {
+  return (
+    <ButtonStyled
+      color={color}
+      variant={variant}
+      size={size}
+      name={name}
+      onClick={onClick}
+    >
+      {children}
+    </ButtonStyled>
+  );
+};
+
+const ButtonStyled = styled(Button)({
+  padding: "11px 15px",
+  fontSize: 12,
+  lineHeight: "14px",
+  borderRadius: 30,
+});
+
+export default CustomButton;
