+++ src/components/Menu/Menu.tsx
@@ -1,11 +1,11 @@
 import {styled} from "@mui/material";
 import Box from "@mui/material/Box";
-import Button from "@mui/material/Button";
 import Link from "@mui/material/Link";
 import React from "react";
 
 import logo from "../../assets/logo.png";
 import {menuItems} from "../../system/constants.ts";
+import CustomButton from "../CustomButton.tsx";
 
 const logoStyle = {
   width: "133px",
@@ -23,7 +23,7 @@ const Menu: React.FC<MenuProps> = ({scrollToSection, onButtonClick}) => {
   return (
     <Box sx={{width: "100%"}}>
       <ButtonsContainer sx={{display: {xs: "none", md: "flex"}}}>
-        <ButtonStyled
+        <CustomButton
           color="secondary"
           variant="outlined"
           size="small"
@@ -31,8 +31,8 @@ const Menu: React.FC<MenuProps> = ({scrollToSection, onButtonClick}) => {
           onClick={onButtonClick}
         >
           Sign up
-        </ButtonStyled>
-        <ButtonStyled
+        </CustomButton>
+        <CustomButton
           color="primary"
           variant="contained"
           size="small"
@@ -40,7 +40,7 @@ const Menu: React.FC<MenuProps> = ({scrollToSection, onButtonClick}) => {
           onClick={onButtonClick}
         >
           Sign in
-        </ButtonStyled>
+        </CustomButton>
       </ButtonsContainer>
       <Box
         sx={{
@@ -86,13 +86,6 @@ const MenuContainer = styled(Box)({
   justifyContent: "end",
   gap: 2,
   paddingBottom: 18,
-});
-
-const ButtonStyled = styled(Button)({
-  padding: "11px 15px",
-  fontSize: 12,
-  lineHeight: "14px",
-  borderRadius: 30,
 });
 
 export default Menu;
