+++ src/sections/HowItWorks.tsx
